requirements.txt
langchain==0.1.4
langchain_openai==0.0.5
openai==1.10.0
python-dotenv==1.0.1
streamlit==1.26.0
